scaler/requirements.txt
Flask
ipaddress
requests
pyOpenSSL==16.2.0
azure-identity
azure-storage-queue
